Add On-Premises Kubernetes Deployment Guide and update README.md
Introduced a new ONPREMISES.md file detailing the deployment of GLAuth on on-premises Kubernetes clusters using NodePort and MetalLB. Updated README.md to include information on on-premises configurations, access methods, and testing instructions. Adjusted values.yaml for NodePort service type and removed ingress template as it's not applicable for on-premises setups.